API Reference

SystemBondGraphs.BondGraph — Method

Create an empty BondGraph to be populated during the analysis

SystemBondGraphs._generate_model — Method

Traverse the BondGraph Structure to create an ODESystem

SystemBondGraphs.add_C! — Method

Create a Nonlinear C-Element with $e = \phi_c(e, f, q, t, ps)$. Setting Causality to true represents the elements being in derivative causality.

SystemBondGraphs.add_C! — Method

Create a Linear C-Element for analysis. Setting Causality to true represents the elements being in derivative causality.

SystemBondGraphs.add_Ce! — Method

Create a "Linear" Concentration-Element for analysis. Setting Causality to true represents the elements being in derivative causality.

SystemBondGraphs.add_Chemostat! — Method

Create a nonliner chemostat

SystemBondGraphs.add_Chemostat! — Method

Create a linear chemostat

SystemBondGraphs.add_Flowstat! — Method

Create a nonlinear Flowstat

SystemBondGraphs.add_Flowstat! — Method

Create a linear Flowstat

SystemBondGraphs.add_GY! — Method

Add a linear gyrator with modulus r, in element name in, out element name out, and named name.

SystemBondGraphs.add_I! — Method

Create a Nonlinear I-Element with $f = \phi_I(e, f, p, t, ps)$. Setting Causality to true represents the elements being in derivative causality.

SystemBondGraphs.add_I! — Method

Create a Linear I-Element, setting the casuality to true signifies that the element is in derivative causality

SystemBondGraphs.add_M! — Method

Create a Nonlinear M-Element with $p = \phi_M(e, f, p, q, t, ps)$. Setting Causality to true represents the elements being in derivative causality.

SystemBondGraphs.add_M! — Method

Create a Linear M-element with causality being set to false. Derivative causality for M-elements is still under development.

SystemBondGraphs.add_MGY! — Method

Add a modulated gyrator with modulus r, in element name in, out element name out, and named name.

SystemBondGraphs.add_MTF! — Method

Add a modulated transformer with modulus m, in element name in, out element name out, and named name.

SystemBondGraphs.add_R! — Method

Create a nonlinear R-Element with $\Phi_r$ registered with modelingtoolkit.jl to prevent simplification through the nonlinear function. $e = \Phi_r(e, f, t, ps)$. Params are for any parameters to the nonlinear function.

SystemBondGraphs.add_R! — Method

Create a Linear R-Element for the bondgraph. Causality will always be false for an R-element since it does not have a "preferred" causality.

SystemBondGraphs.add_Re! — Method

Add a reaction component between species "in" and "out" with parameters R, T, r

SystemBondGraphs.add_Se! — Method

Create a nonlinear effort input with $e = S_e(e, f, iv, params)$ with name

SystemBondGraphs.add_Se! — Method

Create a Symbolic/Constant Effort Input. Creates a system with parameters Se for the ODESystem.

SystemBondGraphs.add_Sf! — Method

Create a nonlinear flow input with $f = S_f(e, f, iv, params)$ with name

SystemBondGraphs.add_Sf! — Method

Create a Symbolic/Constant Flow Input. Creates a system with parameters Sf for the ODESystem.

SystemBondGraphs.add_TF! — Method

Add a linear transformer with modulus m, in element name in, out element name out, and named name.

SystemBondGraphs.generate_model — Method

Generate an ODE System from the BondGraph Structure