BondGraphs.add_1J! — Method

Add a 1-Junction to the bond graph, BG. The elements are specified with a dict of names and directions. true represents the element going "in" to the 1-junction. false represents the element going "out" from the 1-junction

BondGraphs.add_0J! — Method

Add a 0-Junction to the bond graph, BG. The elements are specified with a dict of names and directions. true represents the element going "in" to the 0-junction. false represents the element going "out" from the 0-junction