BondGraphs.add_Bond! — Method

Create an Bond in a BondGraph to connect Junction->Junction, OnePort->TwoPort. It creates an empty bond to be a connector between elements.

BondGraphs.add_R! — Method

Create a Linear R-Element for the bondgraph. Causality will always be false for an R-element since it does not have a "preferred" causality.

BondGraphs.add_R! — Method

Create a nonlinear R-Element with $\Phi_r$ registered with modelingtoolkit.jl to prevent simplification through the nonlinear function. $e = \Phi_r(e, f, t, ps)$. Params are for any parameters to the nonlinear function.

BondGraphs.add_C! — Method

Create a Linear C-Element for analysis. Setting Causality to true represents the elements being in derivative causality.

BondGraphs.add_C! — Method

Create a Nonlinear C-Element with $e = \phi_c(e, f, q, t, ps)$. Setting Causality to true represents the elements being in derivative causality.

BondGraphs.add_I! — Method

Create a Linear I-Element, setting the casuality to true signifies that the element is in derivative causality

BondGraphs.add_I! — Method

Create a Nonlinear I-Element with $f = \phi_I(e, f, p, t, ps)$. Setting Causality to true represents the elements being in derivative causality.

BondGraphs.add_M! — Method

Create a Linear M-element with causality being set to false. Derivative causality for M-elements is still under development.

BondGraphs.add_M! — Method

Create a Nonlinear M-Element with $p = \phi_M(e, f, p, q, t, ps)$. Setting Causality to true represents the elements being in derivative causality.