BondGraphs.add_TF! — Method

Add a linear transformer with modulus m, in element name in, out element name out, and named name.

BondGraphs.add_GY! — Method

Add a linear gyrator with modulus r, in element name in, out element name out, and named name.

BondGraphs.add_MTF! — Method

Add a modulated transformer with modulus m, in element name in, out element name out, and named name.

BondGraphs.add_MGY! — Method

Add a modulated gyrator with modulus r, in element name in, out element name out, and named name.